requirements.txt
pytube3
moviepy
imageio